:hammer: Refactor: 모달 트랜지션 로직과 UI 분리
- 마운트/등장·퇴장 트랜지션 로직을 usePresence 훅으로 추출
- 백드롭·ESC·스크롤잠금·닫기버튼을 담는 범용 Modal 컴포넌트 분리
- LoginModal은 로그인 내용만 담도록 간소화
