+++ src/shared/ui/index.ts
@@ -1,1 +1,2 @@
 export { CloseButton } from './CloseButton';
+export { Modal } from './Modal';
